added stepper library
created the stepper library where its initialized, the velocity can be changed by computing the new prescalar and periods of the timer. also added possibility to change the microstep.
Needs to be implemented in the main.c code.
In the future, position control might be added in case its needed.